requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
python-telegram-bot==21.7
pydantic==2.9.2
pydantic-settings==2.11.0
gspread==6.1.4
google-cloud-vision==3.6.0
